requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
google-generativeai==0.8.5
python-multipart==0.0.6
python-dotenv==1.0.0
websocket-client-asyncio==1.1.3
aiohttp==3.9.1
pydantic==2.5.2
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
logging-config==1.0.3
